+++ src/components/users/SignIn/LocalLoginContainer.jsx
@@ -8,30 +8,7 @@ import { useRecoilState } from "recoil";
 import { userState } from "../../../authentication/userState";
 import * as Sentry from "@sentry/react";
 import { getAccessToken, getRefreshToken } from "../../../authentication/auth";
-const CustomTextField = styled((props) => (
-  <TextField InputProps={{ disableUnderline: true }} {...props} />
-))(({ theme }) => ({
-  "& .MuiFilledInput-root": {
-    border: "none",
-    width: "400px",
-    overflow: "hidden",
-    marginBottom: "15px",
-    borderRadius: 8,
-    backgroundColor: "#edeef0",
-    transition: theme.transitions.create([
-      "border-color",
-      "background-color",
-      "box-shadow",
-    ]),
-    "&:hover": {
-      backgroundColor: "transparent",
-    },
-    "&.Mui-focused": {
-      borderColor: "#3399ff",
-      outline: "2px solid#80bfff",
-    },
-  },
-}));
+import CustomTextField from "../../common/CustomTextField";
 
 function LocalLoginContainer() {
   const [inputs, setInputs] = useState({
