+++ src/components/users/SignIn/styled.js
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import { Link } from "react-router-dom";
 
 const LocalLoginWrapper = styled.div`
   margin-top: 25px;
@@ -53,4 +54,93 @@ const LoginButton = styled.button`
   }
 `;
 
-export { LocalLoginWrapper, LoginInputContainer, LoginButton, Input };
+const AuthMenuContainer = styled.div`
+  margin-top: 30px;
+  margin-left: 80px;
+  margin-right: 80px;
+  position: relative;
+  font-family: "Inter";
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  p {
+    color: #000000cc;
+  }
+`;
+
+const ResetPasswordLink = styled(Link)`
+  text-decoration-line: none;
+  color: #000000cc;
+  &:link {
+    text-decoration-line: none;
+    color: #000000cc;
+  }
+  &:visited {
+    text-decoration-line: none;
+    color: #000000cc;
+  }
+  &:hover {
+    text-decoration-line: none;
+    color: #0064ff;
+  }
+  &:active {
+    text-decoration-line: none;
+    color: #000000cc;
+  }
+`;
+
+const SignUpLink = styled(Link)`
+  position: absolute;
+  top: 0%;
+  right: 0%;
+  text-decoration: none;
+  color: #000000cc;
+  &:link {
+    text-decoration-line: none;
+    color: #000000cc;
+  }
+  &:visited {
+    text-decoration-line: none;
+    color: #000000cc;
+  }
+  &:hover {
+    text-decoration: none;
+    color: #0064ff;
+  }
+  &:active {
+    text-decoration-line: none;
+    color: #000000cc;
+  }
+`;
+const LoginFormWrapper = styled.div`
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+`;
+const SocialLoginWrapper = styled.div`
+  margin-top: 10px;
+  text-align: center;
+`;
+const KaKaoLoginImg = styled.img`
+  width: 400px;
+  height: 50px;
+  padding: 0;
+  border-radius: 10px;
+  border: none;
+
+  object-fit: cover;
+`;
+
+export {
+  LocalLoginWrapper,
+  LoginInputContainer,
+  LoginButton,
+  Input,
+  AuthMenuContainer,
+  ResetPasswordLink,
+  SignUpLink,
+  LoginFormWrapper,
+  SocialLoginWrapper,
+  KaKaoLoginImg,
+};
